backend/requirements.txt
langchain
langchain-community
langchain-google-genai
google-genai
python-dotenv
duckduckgo-search
